backend/requirements.txt
# Kuwait Social AI - Backend Requirements

# Core Framework
Flask==3.0.0  # Updated for security patches
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter[redis]==3.5.0
redis==5.0.1  # For caching

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23  # Updated for security patches
alembic==1.13.1

# Validation & Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0

# Authentication & Security
werkzeug==3.0.1  # Critical security update
cryptography==41.0.7  # Updated for security patches
PyJWT==2.8.0

# Redis & Caching
redis==5.0.1
celery==5.3.4  # Updated for bug fixes

# AI & Content Generation
openai>=1.0.0
Pillow==10.1.0  # Security patches
opencv-python==4.8.1.78

# Language Processing
deep-translator==1.11.4
langdetect==1.0.9
arabic-reshaper==3.0.0
python-bidi==0.4.2

# Security & Sanitization
bleach==6.1.0  # For secure HTML sanitization
tinycss2==1.2.1  # CSS parsing for bleach
zxcvbn==4.4.28  # For advanced password strength assessment

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27

# Social Media APIs
requests==2.31.0
requests-oauthlib==1.3.1

# Telegram Bot
python-telegram-bot==20.5

# Email
Flask-Mail==0.9.1

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk[flask]==1.39.1  # Updated for compatibility
psutil==5.9.6  # System monitoring for performance checks
APScheduler==3.10.4  # Background task scheduling

# Date & Time (Kuwait timezone)
pytz==2023.3
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0  # Updated
pytest-cov==4.1.0
black==23.11.0  # Updated
flake8==6.1.0
click==8.1.7  # CLI framework for management commands
tabulate==0.9.0  # Table formatting for CLI output

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Utilities
python-slugify==8.0.1
validators==0.22.0
phonenumbers==8.13.22
cachetools==5.3.1
urllib3<2.0  # Pin to version compatible with LibreSSL